Temporarily undid LoginAction refactoring to attempt to resolve merge errors
