+++ src/test/java/LoginActionTest.java
@@ -1,4 +1,4 @@
-import UseCases.LoginAction.LoginAction;
+import UseCases.LoginAction;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
